Added README.optimization
Synchronized files/package names in debian/. However, using George Danchev's
debian-addons-bg/bobcat/trunk/debian the whole debian directory might turn out
to be superfluous. For now it's kept.

Defined proper DOC installation directories in INSTALL.im

Added sourcetar to create either a debian-source tar or a generic source tar
which is signed in a detached signature. The lintian overrides appear to be
superfluous and were removed from the install-script. 

For the rest: see the changelog file.



git-svn-id: https://bobcat.svn.sourceforge.net/svnroot/bobcat/trunk@119 95c995b8-3b1c-0410-acf1-adacd5d40d98